Use .env file for environment variables
Adds a .development.env file with developer defaults. This provides one
common place to set defaults (used both locally and in Docker) and
should stop people having environment issues when new variables are
added to runner.

Also introduces a Makefile for running the various commands - less typing
and tab completion on compatible shells.
