--- .coveragerc
@@ -1,3 +1,0 @@
-[run]
-omit = app/templates/*.html,app/dev_mode/templates/*.html
-
