+++ .development.env
@@ -1,0 +1,17 @@
+FLASK_APP=application.py
+FLASK_ENV=development
+EQ_DEVELOPER_LOGGING=True
+EQ_SUBMITTED_RESPONSES_TABLE_NAME=dev-submitted-responses
+EQ_QUESTIONNAIRE_STATE_TABLE_NAME=dev-questionnaire-state
+EQ_SESSION_TABLE_NAME=dev-eq-session
+EQ_USED_JTI_CLAIM_TABLE_NAME=dev-used-jti-claim
+EQ_ENABLE_SECURE_SESSION_COOKIE=False
+EQ_ENABLE_LIVE_RELOAD=True
+EQ_SUBMISSION_BACKEND=log
+EQ_DATASTORE_EMULATOR_CREDENTIALS=True
+DATASTORE_EMULATOR_HOST=localhost:8432
+DATASTORE_DATASET=local
+EQ_REDIS_HOST=localhost
+EQ_REDIS_PORT=6379
+EQ_SECRETS_FILE=dev-secrets.yml
+EQ_KEYS_FILE=dev-keys.yml
