--- jenkins-shred.sh
@@ -1,12 +1,0 @@
-#!/bin/bash
-
-shred -u -v secure-data.tar
-
-shred -u -v secure-data/*
-rm -rf secure-data
-
-shred -u -v $Keys
-rm -f $Keys
-
-shred -u -v secrets.yml
-shred -u -v keys.yml
