--- jwt-test-keys/sdc-sdx-submission-encryption-public-v1.pem
@@ -1,9 +1,0 @@
------BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAu4UO1frTr8LPyyKXkQNi
-ua4VrCeLec/sMkkYt7aFf/jMr2pTCNeyGGD0lp1JZp5u4fs3yVMoQTxfmswggYG/
-44pzNYK7+03GyvGbNrHcLXZ1nPETPhMPWobWOWlQ0YH9vLCLD4nuFUS7ma5qTiRX
-y/UKPPvs5MSYj5EOFqkqziToTzsQ9LJewsLODSvQuILcX9Ygb0seq9zxesEoOPgz
-0xhxht8kTIp24UMhax1QAwOtqjkiShUOnj8xnXjGoc1GOJSSLDGm9dxaOt1F4uC5
-ddIEdYRuxd9A7yWjP1I8jWC/DSq8UEAooVfe6XNsoWWF2zn9XDjhWw0hwEXSX/QT
-9wIDAQAB
------END PUBLIC KEY-----
