--- jwt-test-keys/sdc-sr-authentication-encryption-private-v1.pem
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEogIBAAKCAQEAt8LZnIhuOdL/BC029GOaJkVUAqgp2PcmbFr2Qwhf/514DUUQ
-9sKJ1rvwvbmmW2zE8JRtdY3ey0RXGtMn5UZHs8NReHzMxvsmHN4VuaGEnFmPwO82
-1Tkvg0LpKsLkotcw793FD/fut44N2lhpTSW2Sc82uG0p9A+Kud8HCIaWaluosghk
-9rbMGYDzZQk8cA91GtKJRmIOED4PorB/dexDf37qhuWNQgzyNyTti1DTDUIWyzQQ
-Jp926vLbkOip6Fc2R13hOFNETe68Rrw/h3hXEFS17uPFZHsxvm9PFXX9KZMS25oh
-qbNh97I94LL4o4wybl6LaE6lJEHiD6docD0B6wIDAQABAoIBADRQGyUtzbtWbtTg
-jlF6fWrBt83d07P5gA2+w3aHztR1HrUgYVjuPtFLxZgtXseOKm6WwqaBkdhqByYS
-0Lu7izQIuYvEc4N+Afab3rFu4tKjyIHTV9fRpM4IYVqUCwS0oDDZAH2wRlwo65aq
-LqgQwVk3zUspgJUDS6nobRcnQXDbVaQ54JU0zSXrFJqZygrUR5TDuPnE7Ehbb9Ru
-L1YNkxn2wVT9iOHdyaxr9co7x1z01hHCgdf3SUyGTCOCqp9rJYXtm+GPpZMRpwv7
-CdsMfDxpkNKC2X/hBHz5ux9sC8kRA/JcTKGvbKbPVpedWyIYwKjJ8H1A0zuSQX9Q
-rZU1a0kCgYEA3EyNsBwtllLSzA2sb1yf6qc+0mvXcPPrTt7WWG7bUI1/43Nx2rMW
-XxRGdJlOdA9GnxIK2Ir75pPXj6VVL2U2CCh87td5bnVDr8DMA8nj7efpjMpAUEtU
-QX/qKHtzkr3nRjLLkrL9IhQ6m9rNVtyKqWLTnBv6Uflq2UlYHh2xBi0CgYEA1Yp3
-DycqKDkkGHi9wuZeyAhQDsKHyVEukS5bflVjnkbkdX7Z4efMzKdFknNk6O/YtgYh
-Ti/XheojCkCVMSEJ3kndsotIsEf1kXYIvfSSBPO0J8GWma7meGbUn61Tq8Kj10LI
-8k6KsXiT67+r79wOYcRclIBGNm3nR4rMMpKAj3cCgYAB6oCI+ZXD6vB+adgIF+wk
-JFQ9jEaRau2u/+0pU72Ak9p65fQljM0zAoAiX3r5M3DPzV5ex8atGLgVPcDh6qVv
-qLp9cU5TEZ4HF0wu9ECRPyUe3lt011LiRvSIaZp1ukUarTJsEjZ1Z2ujE2IZ0U07
-b+qbPvsMX3j4btTfXi69+QKBgFZvAHgKsz6quliJbs3X719qNfVzegDbskyjhfch
-2vuy2EBSwyB0ceoYfsmjmaHLi11KJ+r85HDY76vzri+/nr3yCiF9zUNFLTnem/U/
-bGdCuZYp/qpgJ/tuK/wh7S8lzqmP58RkVDE3jDAtWgvxd4TNNWgKb+ESJT5JCRQj
-RpRLAoGALFlPzTd/ifWCcV0Pn5U/ESzX1az2kfll30VxjfGVFsRJ4ZvezxgI+vwo
-OZrki4MBTK/6GFkHLFkF6w2Le+Y5Nos9O2UUZs45lwLEYbQ4yKcx2KlWGLZOypB8
-i7/6TB95Ej2i5KgaSlcJjOyOx7g20TwDD1THtLXgY54d0Yr9T/U=
------END RSA PRIVATE KEY-----
