--- jwt-test-secrets/eq-secret-key.txt
@@ -1,1 +1,0 @@
-SuperSecretDeveloperKey
