--- jwt-test-secrets/eq-server-side-storage-encryption-user-pepper.txt
@@ -1,1 +1,0 @@
-boba.fett.ig88.xizor
