--- jwt-test-secrets/eq-server-side-storage-user-id-salt.txt
@@ -1,1 +1,0 @@
-luke.skywalker.r2d2.c3p0
