--- jwt-test-secrets/eq-server-side-storage-user-ik-salt.txt
@@ -1,1 +1,0 @@
-jabba.leia.organa.solo
