+++ tests/data/jwt-test-keys/sdc-sr-authentication-encryption-public-v1.pem
